renamed home and about pages files, renamed components folder
