+++ src/components/ErrorBoundary/styles.module.scss
