+++ src/routes/routesList.js
@@ -1,9 +1,8 @@
 import { lazy } from "react";
 import { Navigate } from "react-router-dom";
 
-const Home = lazy(() => import("pages/Home"));
-const About = lazy(() => import("pages/About"));
-
+const Home = lazy(() => import("pages/Home/Home"));
+const About = lazy(() => import("pages/About/About"));
 
 const routes = [
   {
